pyproject.toml
[tool.poetry]
name = "openlane"
version = "2.1.8"
description = "An infrastructure for implementing chip design flows"
authors = ["Efabless Corporation and Contributors <donn@efabless.com>"]
readme = "Readme.md"
license = "Apache-2.0"
homepage = "https://efabless.com"
repository = "https://github.com/efabless/openlane"
documentation = "https://openlane.readthedocs.io"


[tool.poetry.dependencies]
python = ">=3.8,<4"
click = ">=8,<9"
cloup = ">=3.0.5,<4"
pyyaml = ">=5,<7"
rich = ">=12,<14"
volare = ">=0.16.0"
lxml = ">=4.9.0"
deprecated = ">=1.2.10,<2"
libparse = ">=0.3.1,<1"
psutil = ">=5.9.0"
httpx = ">=0.22.0,<0.28"
klayout = ">=0.29.0,<0.30.0"
rapidfuzz = ">=3.9.0,<4"
ioplace-parser = "0.3.0"
yamlcore = "^0.0.2"


[tool.poetry.dev-dependencies]
wheel = "*"
black = ">=24.4.0,<25"

flake8 = ">=4"
flake8-no-implicit-concat = "0.3.3"
flake8-pytest-style = "*"

mypy = "*"
lxml-stubs = "*"
types-urllib3 = "*"
types-typed-ast = "*"
types-six = "*"
types-setuptools = "*"
types-PyYAML = "*"
types-docutils = "*"
types-decorator = "*"
types-commonmark = "*"
types-colorama = "*"
types-Pygments = "*"
types-Deprecated = "*"
types-psutil = "*"

pytest = "*"
pytest-xdist = "*"
pytest-cov = "*"
pyfakefs = ">=5.2.3,<6"
pillow = ">=10.0.1,<11"


[tool.poetry.scripts]
openlane = "openlane.__main__:cli"
"openlane.steps" = "openlane.steps.__main__:cli"
"openlane.config" = "openlane.config.__main__:cli"
"openlane.state" = "openlane.state.__main__:cli"
"openlane.env_info" = "openlane:env_info_cli"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
